+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Load data from Dgraph cluster as [GraphFrames](https://graphframes.github.io/graphframes/docs/_site/index.html) `GraphFrame`.
+- Use exact uid cardinality for uid range partitioning. Combined with predicate partitioning, large
+  predicates get split into more partitions than small predicates.
 
 ### Fixed
 - Dgraph groups with no predicates caused a `NullPointerException`.